In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -3,13 +3,13 @@
 /**
  * main - entry point
  * description: prints the first 50 Fibonacci numbers
- * return: 0
+ * Return: always 0
  */
 
 int main(void)
 {
 	int a;
-	int fib[50];
+	long int fib[50];
 
 	fib[0] = 1;
 	fib[1] = 2;
@@ -20,7 +20,7 @@ int main(void)
 	}
 	for (a = 0; a < 50; a++)
 	{
-		printf("%d", fib[a]);
+		printf("%ld", fib[a]);
 		if (a != 49)
 		{
 			printf(", ");
